requirements.txt
-e git+https://github.com/uma-pi1/kge.git@308cebd21e58119d05b1c8613c195cdafe0a112b#egg=libkge
numpy==1.19.5
ogb==1.3.2
pandas==1.3.2
pytorch-lightning==1.4.6
PyYAML==5.1.2
scikit-learn==0.24.2
scipy==1.7.1
tokenizers==0.10.3
torch==1.7.1
tqdm==4.62.2
transformers==4.10.2
